backend/requirements.txt
fastapi>=0.118.0
uvicorn[standard]>=0.32.0
pydantic>=2.11.0
python-dotenv>=1.0.1
anthropic>=0.40.0
httpx>=0.28.0
pandas>=2.2.3
numpy>=2.2.0
pytest>=8.3.3
pytest-asyncio>=0.24.0